added buildCoverageRegex and buildGitStrategy to Project
