+++ src/main/java/org/gitlab4j/api/Constants.java
@@ -825,5 +825,28 @@ public interface Constants {
             return (enumHelper.toString(this));
         }
     }
+
+    /** Enum for the build_git_strategy of the project instance. */
+    enum BuildGitStrategy {
+
+        FETCH, CLONE;
+
+        private static JacksonJsonEnumHelper<BuildGitStrategy> enumHelper = new JacksonJsonEnumHelper<>(BuildGitStrategy.class);
+
+        @JsonCreator
+        public static BuildGitStrategy forValue(String value) {
+            return enumHelper.forValue(value);
+        }
+
+        @JsonValue
+        public String toValue() {
+            return (enumHelper.toString(this));
+        }
+
+        @Override
+        public String toString() {
+            return (enumHelper.toString(this));
+        }
+    }
 }
